requirements.txt
# Azure ML SDK v2 (the modern SDK)
azure-ai-ml>=1.12.0
azure-identity>=1.15.0

# ML Libraries
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0

# MLflow for experiment tracking
mlflow>=2.10.0
azureml-mlflow>=1.55.0

# Visualization
matplotlib>=3.8.0

# Responsible AI
raiwidgets>=0.34.0
responsibleai>=0.34.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
